requirements.txt
transformers==4.32.0
sentencepiece
tensorboard
datasets
accelerate
bitsandbytes
peft
scipy
protobuf==3.20
numpy==1.23
scikit-learn
pillow
opencv-python
opencv-contrib-python
pydantic
fastapi
fake_headers
json_repair
uvicorn
redis
redisvl
beautifulsoup4
asyncio
fastapi_limiter
vllm
qwen-vl-utils
flash-attn
optimum
